dynamic programming and shared preferences
everything is done dynamically (code is run through java. XML is only used for id and creating the activities themselves), there are a few buttons, one which randomly selects and puts an icon/image on the screen(there are 6 images/icons), another button which randomly selects a color for the background and changes the background to the color. two dialog's- one which is called settings which tells contains a text of the current time, and a login dialog using shared preferences which takes a username and password and is only saved when on the main activity the save&exit button is pressed which also exits the application.